Fix Talk-mode crash: babl VM pointer truncation + dropped args + no-CNV-record guard
Fixes the crash in bug-critter-talk.txt (right-click Talk on an NPC with
no CNV.ARK conversation record, e.g. Bragit): EXC_BAD_ACCESS deep in the
babl conversation VM.

Root causes, in the order they were hit:
- read_archive_entry() called with a dropped 3rd (destination) argument,
  silently corrupting every NPC's conversation load, not just Bragit's.
- FUN_00019470's "no record for this NPC" path hardcoded `return 1`
  instead of forwarding message_scroll_print_wrapped()'s real return
  value, so the caller's `sVar1 < 0` success check always passed even
  on this path.
- Two more dropped-argument sites feeding message_scroll_print_wrapped
  its message string (register-forwarding in the original ARM code that
  this host's codegen doesn't reproduce for calls it only sees 1 arg
  for).
- The babl VM's own free-list heap allocator (FUN_00018ac8/ccc/f34)
  packed its 32-bit-only pointer links as 4 individual bytes in the
  block header -- meaningless once real 64-bit pointers flow through
  it. Replaced with the host's real malloc/free (resize is a no-op;
  its own call site ignores the returned pointer).
- DAT_000bbf70/DAT_000bbf00/DAT_000bbf14/DAT_000bbf0c (babl symbol
  table base pointers) were declared `int`, truncating the real 64-bit
  heap pointers babl_alloc() now hands back. Widened to intptr_t.
- With no CNV record loaded, DAT_000bbf70 stays 0 (never initialized),
  but FUN_00028c00's success branch dereferenced it unconditionally
  anyway; added `DAT_000bbf70 != 0` guards across the whole
  babl-symbol-table reader cluster and gated FUN_00028c00's own
  success branch on it.

Confirmed via lldb against this port's own compiled binary (not the
original ARM code) for each crash site. demo_critter_talk_test.txt now
runs to completion; full demo_*.txt regression suite (6/6, excluding
the slow automap script) still clean.

+++ uw.h
@@ -542,7 +542,7 @@ void FUN_0001853c();
 void FUN_000188fc();
 uint *FUN_00018ac8();
 void FUN_00018ccc();
-int FUN_00018f34();
+intptr_t FUN_00018f34();
 undefined4 FUN_00019120();
 void FUN_0001927c();
 undefined4 FUN_00019470();
